Refactoring search observer and adding refresh tweet use case

+++ repo/src/main/java/com/example/repo/search/model/SearchMetadata.java
@@ -23,7 +23,7 @@ public class SearchMetadata {
     public long maxId;
 
     @SerializedName("since_id")
-    public int sinceId;
+    public long sinceId;
 
     @SerializedName("completed_in")
     public double completedIn;
